requirements.txt
asgiref==3.5.0
Django==3.2.12
django-crispy-forms==1.14.0
Faker==13.3.1 
factory-boy==3.2.1
pytz==2021.3
sqlparse==0.4.2

django-heroku==0.3.1
gunicorn==20.1.0

